Revert "Restore & obsolete PageRenderer"
This reverts commit 11cedb868af3a3a16993765a768ccaba9cabc4ce.
