+++ Xamarin.Forms.Platform.iOS/Renderers/CarouselPageRenderer.cs
@@ -253,7 +253,7 @@ namespace Xamarin.Forms.Platform.iOS
 			UIView container = new CarouselPageContainer(page);
 
 			UIView view = renderer.NativeView;
-			if (renderer is AccessiblePageRenderer) // AccessiblePageRenderer's NativeView is the PageContainer, but we need the parent View.
+			if (renderer is PageRenderer) // PageRenderer's NativeView is the PageContainer, but we need the parent View.
 				view = view.Superview;
 
 			container.AddSubview(view);
