+++ Xamarin.Forms.Platform.iOS/Renderers/PageRenderer.cs
@@ -8,9 +8,7 @@ using PageUIStatusBarAnimation = Xamarin.Forms.PlatformConfiguration.iOSSpecific
 
 namespace Xamarin.Forms.Platform.iOS
 {
-	[Obsolete("PageRenderer is obsolete as of 4.0.0. Please use AccessiblePageRenderer instead.")]
-	[EditorBrowsable(EditorBrowsableState.Never)]
-	public class PageRenderer : UIViewController, IVisualElementRenderer, IEffectControlProvider
+	public class PageRenderer : AccessibleUIViewController, IVisualElementRenderer, IEffectControlProvider
 	{
 		bool _appeared;
 		bool _disposed;
@@ -29,7 +27,7 @@ namespace Xamarin.Forms.Platform.iOS
 
 		void IEffectControlProvider.RegisterEffect(Effect effect)
 		{
-			VisualElementRenderer<VisualElement>.RegisterEffect(effect, View);
+			VisualElementRenderer<VisualElement>.RegisterEffect(effect, NativeView);
 		}
 
 		public VisualElement Element { get; private set; }
@@ -43,10 +41,10 @@ namespace Xamarin.Forms.Platform.iOS
 
 		public UIView NativeView
 		{
-			get { return _disposed ? null : View; }
+			get { return _disposed ? null : Container; }
 		}
 
-		public void SetElement(VisualElement element)
+		public override void SetElement(VisualElement element)
 		{
 			VisualElement oldElement = Element;
 			Element = element;
@@ -54,11 +52,15 @@ namespace Xamarin.Forms.Platform.iOS
 
 			OnElementChanged(new VisualElementChangedEventArgs(oldElement, element));
 
-			if (Element != null && !string.IsNullOrEmpty(Element.AutomationId))
-				SetAutomationId(Element.AutomationId);
+			base.SetElement(element);
 
 			if (element != null)
+			{
+				if (!string.IsNullOrEmpty(element.AutomationId))
+					SetAutomationId(element.AutomationId);
+
 				element.SendViewInitialized(NativeView);
+			}
 
 			EffectUtilities.RegisterEffectControlProvider(this, oldElement, element);
 		}
@@ -72,6 +74,9 @@ namespace Xamarin.Forms.Platform.iOS
 		{
 			base.ViewDidLayoutSubviews();
 
+			if (_disposed)
+				return;
+
 			if (Element.Parent is BaseShellItem)
 				Element.Layout(View.Bounds.ToRectangle());
 
@@ -81,16 +86,14 @@ namespace Xamarin.Forms.Platform.iOS
 		public override void ViewSafeAreaInsetsDidChange()
 		{
 			UpdateShellInsetPadding();
-			var page = (Element as Page);
-			if (page != null && Forms.IsiOS11OrNewer)
+			if (Page != null && Forms.IsiOS11OrNewer)
 			{
-				var insets = NativeView.SafeAreaInsets;
-				if (page.Parent is TabbedPage)
+				var insets = View.SafeAreaInsets;
+				if (Page.Parent is TabbedPage)
 				{
 					insets.Bottom = 0;
 				}
-				page.On<PlatformConfiguration.iOS>().SetSafeAreaInsets(new Thickness(insets.Left, insets.Top, insets.Right, insets.Bottom));
-
+				Page.On<PlatformConfiguration.iOS>().SetSafeAreaInsets(new Thickness(insets.Left, insets.Top, insets.Right, insets.Bottom));
 			}
 
 			base.ViewSafeAreaInsetsDidChange();
@@ -133,13 +136,16 @@ namespace Xamarin.Forms.Platform.iOS
 		{
 			base.ViewDidLoad();
 
-			var uiTapGestureRecognizer = new UITapGestureRecognizer(a => View.EndEditing(true));
+			if (NativeView == null)
+				return;
+
+			var uiTapGestureRecognizer = new UITapGestureRecognizer(a => NativeView.EndEditing(true));
 
 			uiTapGestureRecognizer.ShouldRecognizeSimultaneously = (recognizer, gestureRecognizer) => true;
 			uiTapGestureRecognizer.ShouldReceiveTouch = OnShouldReceiveTouch;
 			uiTapGestureRecognizer.DelaysTouchesBegan =
 				uiTapGestureRecognizer.DelaysTouchesEnded = uiTapGestureRecognizer.CancelsTouchesInView = false;
-			View.AddGestureRecognizer(uiTapGestureRecognizer);
+			NativeView.AddGestureRecognizer(uiTapGestureRecognizer);
 
 			UpdateBackground();
 
@@ -150,16 +156,16 @@ namespace Xamarin.Forms.Platform.iOS
 			_tracker = new VisualElementTracker(this, !(Element.Parent is BaseShellItem));
 
 			_events = new EventTracker(this);
-			_events.LoadEvents(View);
+			_events.LoadEvents(NativeView);
 
-			Element.SendViewInitialized(View);
+			Element.SendViewInitialized(NativeView);
 		}
 
 		public override void ViewWillDisappear(bool animated)
 		{
 			base.ViewWillDisappear(animated);
 
-			View.Window?.EndEditing(true);
+			NativeView?.Window?.EndEditing(true);
 		}
 
 		protected override void Dispose(bool disposing)
@@ -262,6 +268,9 @@ namespace Xamarin.Forms.Platform.iOS
 
 		void UpdateShellInsetPadding()
 		{
+			if (Element == null)
+				return;
+
 			var setInsets = Shell.GetSetPaddingInsets(Element);
 
 			if (!setInsets && Element.Parent != null)
@@ -297,7 +306,7 @@ namespace Xamarin.Forms.Platform.iOS
 				UIView.Animate(0.25, () => SetNeedsStatusBarAppearanceUpdate());
 			else
 				SetNeedsStatusBarAppearanceUpdate();
-			View.SetNeedsLayout();
+			NativeView?.SetNeedsLayout();
 		}
 
 		bool OnShouldReceiveTouch(UIGestureRecognizer recognizer, UITouch touch)
@@ -327,17 +336,20 @@ namespace Xamarin.Forms.Platform.iOS
 
 		void UpdateBackground()
 		{
+			if (NativeView == null)
+				return;
+
 			string bgImage = ((Page)Element).BackgroundImage;
 			if (!string.IsNullOrEmpty(bgImage))
 			{
-				View.BackgroundColor = ColorExtensions.FromPatternImageFromBundle(bgImage);
+				NativeView.BackgroundColor = ColorExtensions.FromPatternImageFromBundle(bgImage);
 				return;
 			}
 			Color bgColor = Element.BackgroundColor;
 			if (bgColor.IsDefault)
-				View.BackgroundColor = UIColor.White;
+				NativeView.BackgroundColor = UIColor.White;
 			else
-				View.BackgroundColor = bgColor.ToUIColor();
+				NativeView.BackgroundColor = bgColor.ToUIColor();
 		}
 
 		void UpdateTitle()
